# Conflicts:
#	plugin/src/main/java/fr/euphyllia/skyllia/listeners/bukkitevents/paper/PortalAlternativePaperEvent.java

+++ plugin/src/main/java/fr/euphyllia/skyllia/database/postgresql/PostgreSQLIslandCustomData.java
@@ -92,6 +92,7 @@ public class PostgreSQLIslandCustomData extends IslandCustomDataQuery {
 
     @Override
     public @Nullable <P, C> C get(@NotNull NamespacedKey namespace, @NotNull Island island, @NotNull String dataKey, @NotNull PersistentDataType<P, C> type) {
+        if (!ensureTableExists(namespace)) return null;
         String tableName = getTableName(namespace);
         String sql = String.format("""
                 SELECT data_value FROM %s
@@ -120,6 +121,7 @@ public class PostgreSQLIslandCustomData extends IslandCustomDataQuery {
 
     @Override
     public boolean has(@NotNull NamespacedKey namespace, @NotNull Island island, @NotNull String dataKey) {
+        if (!ensureTableExists(namespace)) return false;
         String tableName = getTableName(namespace);
         String sql = String.format("""
                 SELECT 1 FROM %s
@@ -149,6 +151,7 @@ public class PostgreSQLIslandCustomData extends IslandCustomDataQuery {
 
     @Override
     public @NotNull Set<String> getKeys(@NotNull NamespacedKey namespace, @NotNull Island island) {
+        if (!ensureTableExists(namespace)) return new HashSet<>();
         String tableName = getTableName(namespace);
         String sql = String.format("""
                 SELECT data_key FROM %s
@@ -171,6 +174,7 @@ public class PostgreSQLIslandCustomData extends IslandCustomDataQuery {
 
     @Override
     public boolean remove(@NotNull NamespacedKey namespace, @NotNull Island island, @NotNull String dataKey) {
+        if (!ensureTableExists(namespace)) return false;
         String tableName = getTableName(namespace);
         String sql = String.format("""
                 DELETE FROM %s
@@ -183,6 +187,7 @@ public class PostgreSQLIslandCustomData extends IslandCustomDataQuery {
 
     @Override
     public boolean clear(@NotNull NamespacedKey namespace, @NotNull Island island) {
+        if (!ensureTableExists(namespace)) return false;
         String tableName = getTableName(namespace);
         String sql = String.format("""
                 DELETE FROM %s
@@ -195,6 +200,7 @@ public class PostgreSQLIslandCustomData extends IslandCustomDataQuery {
 
     @Override
     public int size(@NotNull NamespacedKey namespace, @NotNull Island island) {
+        if (!ensureTableExists(namespace)) return 0;
         String tableName = getTableName(namespace);
         String sql = String.format("""
                 SELECT COUNT(*) FROM %s
